+++ packages/atomic/src/components/atomic-pager/atomic-pager.tsx
@@ -19,11 +19,14 @@ import {
   FocusTarget,
   FocusTargetController,
 } from '../../utils/accessibility-utils';
+import {randomID} from '../../utils/utils';
+import {RadioButton} from '../common/radio-button';
 
 /**
  * The `atomic-pager` provides buttons that allow the end user to navigate through the different result pages.
  *
- * @part buttons - The list of buttons.
+ * @part buttons - The list of the next/previous buttons and page-buttons.
+ * @part page-buttons - The list of page buttons.
  * @part previous-button - The previous button.
  * @part next-button - The next button.
  * @part page-button - The page button.
@@ -38,6 +41,7 @@ export class AtomicPager implements InitializableComponent {
   @InitializeBindings() public bindings!: Bindings;
   private pager!: Pager;
   public searchStatus!: SearchStatus;
+  private readonly radioGroupName = randomID('atomic-pager-');
 
   @BindStateToController('pager')
   @State()
@@ -71,50 +75,59 @@ export class AtomicPager implements InitializableComponent {
     this.scrollToTopEvent.emit();
   }
 
+  private selectPage(page: number) {
+    this.pager.selectPage(page);
+    this.activePage.focusAfterSearch().then(() => this.scrollToTop());
+  }
+
   private get previousButton() {
     return (
-      <li>
-        <Button
-          style="outline-primary"
-          ariaLabel={this.bindings.i18n.t('previous')}
-          onClick={() => {
-            this.pager.previousPage();
-            this.scrollToTop();
-          }}
-          part="previous-button"
-          disabled={!this.pagerState.hasPreviousPage}
-        >
-          <atomic-icon
-            icon={ArrowRight}
-            class="w-5 transform rotate-180"
-          ></atomic-icon>
-        </Button>
-      </li>
+      <Button
+        style="outline-primary"
+        ariaLabel={this.bindings.i18n.t('previous')}
+        onClick={() => {
+          this.pager.previousPage();
+          this.scrollToTop();
+        }}
+        part="previous-button"
+        disabled={!this.pagerState.hasPreviousPage}
+      >
+        <atomic-icon
+          icon={ArrowRight}
+          class="w-5 transform rotate-180"
+        ></atomic-icon>
+      </Button>
     );
   }
 
   private get nextButton() {
     return (
-      <li>
-        <Button
-          style="outline-primary"
-          ariaLabel={this.bindings.i18n.t('next')}
-          onClick={() => {
-            this.pager.nextPage();
-            this.scrollToTop();
-          }}
-          part="next-button"
-          disabled={!this.pagerState.hasNextPage}
-        >
-          <atomic-icon icon={ArrowRight} class="w-5"></atomic-icon>
-        </Button>
-      </li>
+      <Button
+        style="outline-primary"
+        ariaLabel={this.bindings.i18n.t('next')}
+        onClick={() => {
+          this.pager.nextPage();
+          this.scrollToTop();
+        }}
+        part="next-button"
+        disabled={!this.pagerState.hasNextPage}
+      >
+        <atomic-icon icon={ArrowRight} class="w-5"></atomic-icon>
+      </Button>
     );
   }
 
   private get pages() {
     const pages = this.pager.state.currentPages;
-    return pages.map((page) => this.buildPage(page));
+    return (
+      <div
+        part="page-buttons"
+        role="radiogroup"
+        class="h-10 flex space-x-2 flex-wrap"
+      >
+        {pages.map((page) => this.buildPage(page))}
+      </div>
+    );
   }
 
   private buildPage(page: number) {
@@ -124,22 +137,19 @@ export class AtomicPager implements InitializableComponent {
       parts.push('active-page-button');
     }
     return (
-      <li>
-        <Button
-          ariaCurrent={isSelected ? 'page' : 'false'}
-          style="outline-neutral"
-          ariaLabel={this.bindings.i18n.t('page-number', {page})}
-          onClick={() => {
-            this.pager.selectPage(page);
-            this.scrollToTop();
-            this.activePage.focusAfterSearch();
-          }}
-          class={`btn-page ${isSelected ? 'selected' : ''}`}
-          part={parts.join(' ')}
-          text={page.toLocaleString(this.bindings.i18n.language)}
-          ref={isSelected ? this.activePage.setTarget : undefined}
-        ></Button>
-      </li>
+      <RadioButton
+        key={page}
+        groupName={this.radioGroupName}
+        style="outline-neutral"
+        checked={isSelected}
+        ariaCurrent={isSelected ? 'page' : 'false'}
+        ariaLabel={this.bindings.i18n.t('page-number', {page})}
+        onChecked={() => this.selectPage(page)}
+        class="btn-page"
+        part={parts.join(' ')}
+        text={page.toLocaleString(this.bindings.i18n.language)}
+        ref={isSelected ? this.activePage.setTarget : undefined}
+      ></RadioButton>
     );
   }
 
@@ -150,11 +160,11 @@ export class AtomicPager implements InitializableComponent {
 
     return (
       <nav aria-label={this.bindings.i18n.t('pagination')}>
-        <ul part="buttons" class="h-10 flex space-x-2 flex-wrap">
+        <div part="buttons" class="h-10 flex space-x-2 flex-wrap">
           {this.previousButton}
           {this.pages}
           {this.nextButton}
-        </ul>
+        </div>
       </nav>
     );
   }